Fix GH-17107: Building ext/intl fails with GCC 8.5.0
A previous fix to be able to build C++ extensions with MSVC[1], was
based on two assumptions:

* `max_align_t` would be defined in stddef.h on Windows
* additionally including cstddef as extern C++ would not interfere with
  the already included stddef.h

The former was plain wrong; there is no such typedef (or macro); the
latter was presumptuous, as can be seen in the bug tracker ticket.

Thus we revert that fix, and instead make an exception for Windows,
where we always use the fallback definition, which should work fine on
Windows.

[1] <https://github.com/php/php-src/commit/ab449a7e466570f5700ff91548d2154e9c447a8e>