Run stream_server_reneg_limit.phpt on Windows, too
As is, the client code may block on Windows when trying to read STDERR;
if that happens, neither the openssl nor the php processes would be
terminated, causing further issues for the test suite.

We fix this by using `stream_select()` and only read if there is data
available.  This may cause some delay until we hit the timeout, but it
might be necessary to "empty" STDERR.