Support alternative color quantization algorithms
As is, ext/gd only supports the default color quantization algorithm,
which is used for `imagetruecolortopalette()`.  However, libgd supports
alternative algorithms as of GD 2.1.0, namely NeuQuant (which is built
in) and libimagequant (which is used by pngquant).

We add support for these alternative color quantization algorithms as
small wrappers around the two respective libgd functions, introducing
also a couple of constants.  We also add the nnquant module to our
bundled libgd.