Implement slow GD test helpers in C
There are two test helpers which need to traverse all pixels of the
given images, what is obviously rather slow in PHP.  To avoid that
performance penalty, we implement the functions in C and make them
available as internal PHP functions, but only if `GD_TEST_HELPERS` is
defined when building.  We still keep the PHP implementations as fall-
back in case someone wants to run the tests without the internal
functions being defined.

We also improve the userland functions by traversing the pixels in Z
order what is more cache efficient for the libgd implementation.