rfc: error_backtraces_v2
This commit adds an INI setting, 'fatal_error_backtraces', which users
can enable or disable to control whether or not PHP produces backtraces
for fatal errors.

It defaults to enabled, meaning that any non-recoverable error will now
have a backtrace associated with it. For example, a script timeout will
now look like:

  Fatal error: Maximum execution time of 1 second exceeded in example.php on line 23
  Stack trace:
  #0 example.php(23): usleep(10000)
  #1 example.php(24): recurse()
  #2 example.php(24): recurse()
  ...

It respects the `zend.exception_ignore_args` INI setting and the
SensitiveParameter attributes, so users can ensure that sensitive
arguments do not end up in the backtrace.