Fix searching for libraries like LDAP
The checks for things like ext/ldap will assume libraries are in ./lib,
and not lib64 like they are on many multilib systems, Gentoo included.

Gross that we're hardcoding it, but we should also probably convert
ldap to pkg-config sometime. That may require getting rid of Oracle or
Solaris LDAP support though.