ext/mysql: Don't mention errors for optional part of the test
If the user we're running the tests as in MySQL is an unprivileged user,
we skip the CREATE DATABASE and CREATE SERVER tests. However, the
cleanup code still assumes they're there, and will fail with instead.

Instead, we simply don't print the messages for the equivalent DROP
calls. However, we still do for TABLE, since if we can't run anything in
the test database as the user we're running as, far more tests will
fail.

I don't know if this is the best approach; silencing the errors for when
those tests do run is unwise.