ext/mysqli: Use tempnam to make tracefiles unique
While a CONFLICTS acts as a barrier currently, it'd be nice to remove
that. In addition, if multiple users run the test on the same system,
it'd be owned by one user if a cleanup isn't run properly. The right
solution for temporary files is tempnam and friends instead to generate
a unique filename per invocation.

mysqli_debug_ini.phpt is not yet treated with this, because it uses an
--INI-- section instead; generating a unique name here is trickier.