Support building a shared php8embed.dll
So far only static php8embed.lib has been supported on Windows, while
on POSIX the choice was up to the user.  Now, `--enable-embed=shared`
is also supported on Windows.

For BC reasons we stick with static as default; the resulting embed
library can be used like before.  To avoid that snapshot builds now
build a shared library, we add an exemption to confutils.js.

To properly support a static and a shared library, respectively, we do
the customary `__declspec` dance.