opcache: Do not disable userland error handlers in `opcache_compile_file()`
The logic to disable userland error handlers existed since the first commit
including OPcache in php-src. The reason unfortunately is not explained. No
existing tests require that userland error handlers do not trigger in
`opcache_compile_file()`. The newly added tests are intended to exercise all
possible kinds of edge-cases that might arise and cause issues (e.g. throwing
Exceptions, performing a bailout, or loading a different file from within the
error handler).

They all pass for both `--repeat 1` and `--repeat 2`, as well as with OPcache
enabled and without OPcache enabled (in the latter case with the exception of
the `opcache_get_status()` verification). The list of cached files at the end
of execution also matches my expectations.

Therefore it seems that disabling the userland error handler when compiling a
file with OPcache is not (or at least no longer) necessary.

Fixes php/php-src#17422.