Fix GH-8038: DOM extension reflection version is incorrect
It doesn't appear to be reasonable to pretend that nothing in the DOM
API would have changed in the last 21 years, and it's also not that
reasonable to pretend that we would thoroughly bump that version in the
future.  Thus we go with the common convention to use the `PHP_VERSION`
for bundled extensions.